ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
groq==0.3.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
